Split vultron/core/behaviors/status/nodes.py into nodes/ subpackage
Splits the 1281-line nodes.py (17 classes) into four domain-focused
submodules following the established case/nodes/ pattern (BTND-07-001,
BTND-07-003, CS-18-001 through CS-18-004):

- nodes/conditions.py     — VerifySenderIsParticipantNode (step 1)
- nodes/broadcast.py      — _find_case_manager_id + 5 broadcast nodes (step 3)
- nodes/participant_status.py — 8 nodes for steps 2–5
- nodes/case_status.py    — CASE_STATUS_ALREADY_PRESENT + 3 case status nodes

nodes/__init__.py re-exports all 19 public names for backward
compatibility; no call-site changes required (AC-2).

Test layout mirrors source split: test/core/behaviors/status/nodes/
with one test_<submodule>.py per submodule, 32 new tests (AC-4).

All submodules are under 500 lines (AC-3). All existing tests pass.

Closes #876
